Fix StartSpanFromContextWithTracer to use parent tracer
When 3rd party libraries use opentracing, they don't have the original
tracer. When they grab the current span from the context, all they have
is the span to work off of. The current behavior is for them to start a
span from the context (which may have a parent span with associated
tracer), but that tracer is ignored and the global tracer is used.

This makes it impossible for applications to use these libraries without
also setting a global tracer. Instead of propogating the global tracer
requirement, I've fixed StartSpanFromContextWithTracer so that when a
parent span is set, the child span should use that tracer.

This allows us to avoid using a global tracer even with 3rd party
libraries that rely on using the context to do tracing.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>